// packages/proxyservice/src/interfaces/CustomerLoginRequest.ts
/**
 * This file was automatically generated by joi-to-typescript
 * Do not modify this file manually
 */

export interface CustomerLoginRequest {
  ActivityId?: string;
  CustomerProfileNumber?: string;
  EmailAddress?: string;
  IsEncrypted?: boolean;
  KioskId: number;
  LanguageCode?: string;
  MessageId: string;
  Password?: string;
  PhoneNumber?: string;
  Pin?: string;
}